move subcode to code folder
