+++ app.js
@@ -2,6 +2,6 @@ const dotenv = require("dotenv").config();
 
 const localDB = process.env.LOCAL_JSONFILE
 
-const uiCli = require('./input');
+const uiCli = require('./code/input');
